requirements.txt

# download from pypi.org or whatever defined in ~/.config/pip/pip.conf
build
pyyaml
pytest
pytest-cov

## download source code from github repo and build it
# urllib3 1.26.8
urllib3 @ https://github.com/urllib3/urllib3/releases/download/1.26.8/urllib3-1.26.8-py2.py3-none-any.whl

# or, use find-links
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0
